requirements.txt
pyradiomics
streamlit
pandas
numpy
matplotlib
lifelines
joblib
scikit-survival
plotly
SimpleITK 
nibabel
scikit-learn
scipy
osqp
